Refactor JointJS graph and paper setup into util file
