+++ src/components/editor/Paper.vue
@@ -6,28 +6,20 @@
 
 <script>
 import joint from 'jointjs'
+import { createPaper } from '../../util/jointjs_util'
 
 export default {
   props: ['graph'],
+  data () {
+    return {
+      paper: null
+    }
+  },
   mounted () {
-    // Store reference to the Vue element
-    const vue = this
     // Pull graph out of props
     const graph = this.graph
-
-    // Define paper
-    const paper = new joint.dia.Paper({
-      el: this.$refs.paper,
-      gridSize: 1,
-      model: graph
-    })
-
-    // When clicking a JointJS element, emit event containing element model
-    paper.on('cell:pointerdown',
-      (cellView) => {
-        vue.$emit('send-element', cellView.model)
-      },
-    )
+    // Set up paper
+    this.paper = createPaper(this.$refs.paper, graph, this)
 
     // Define some basic shapes
     const usersTable = new joint.shapes.basic.Rect({
@@ -56,33 +48,6 @@ export default {
 
     // Add the cells to the graph (model)
     graph.addCells([usersTable, colId, colUsername])
-
-    // Handle bounding child elements inside parent element
-    graph.on('change:position', (cell, newPosition) => {
-      const parentId = cell.get('parent')
-
-      // Move the element if it doesn't have a parent
-      if (!parentId) return
-
-      if (!cell.get('originalPosition')) {
-        // Set originalPosition the first time
-        cell.set('position', cell.previous('position'))
-        cell.set('originalPosition', cell.position({ parentRelative: true }))
-        cell.set('position', newPosition)
-      }
-
-      // Calculate relative offsets
-      let relativePos = cell.position({ parentRelative: true })
-      let offsetY = cell.position({ parentRelative: true }).y - cell.get('originalPosition').y
-
-      // Reset child position if its relative position has changed
-      if (offsetY || relativePos.x) {
-        cell.set('position', cell.previous('position'))
-      }
-
-      // Update originalPosition
-      cell.set('originalPosition', cell.position({ parentRelative: true }))
-    })
   }
 }
 </script>
